% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{construct_data}
\alias{construct_data}
\title{Construct data}
\usage{
construct_data(p = 17)
}
\arguments{
\item{p}{number of predictors. Default is 17.}
}
\value{
tibble with columns Deltager, Uge and simulated answers for predictors 1:p
}
\description{
Construct data
}
