% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_excel.R
\name{download_excel}
\alias{download_excel}
\title{Download excel}
\usage{
download_excel(data, file_name)
}
\arguments{
\item{data}{data to download}

\item{file_name}{file name including extension. Should be .xlsx}
}
\value{
excel file
}
\description{
Download excel
}
