% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plots_as_list.R
\name{get_file_name_from_patient_and_question}
\alias{get_file_name_from_patient_and_question}
\title{Get file name from patient and question}
\usage{
get_file_name_from_patient_and_question(str_patient, str_question)
}
\description{
Get file name from patient and question
}
