% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_question.R
\name{get_plot_data}
\alias{get_plot_data}
\title{Get plot data}
\usage{
get_plot_data(data, var_names)
}
\value{
data with summarised questions grouped based on var_names
}
\description{
This is a helper function for plot_question
It takes user specified data and returns summarised output
}
