% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot_file_name.R
\name{get_plot_file_name}
\alias{get_plot_file_name}
\title{Get plot file name}
\usage{
get_plot_file_name(patient, question, filetype = "png")
}
\arguments{
\item{patient}{character string with patient ID}

\item{question}{character string with question (see get_questions())}

\item{filetype}{character string with filetype (default is png)}
}
\description{
Get plot file name
}
