% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_random_data}
\alias{get_random_data}
\title{Get random data}
\usage{
get_random_data(data_pmap, p)
}
\arguments{
\item{data_pmap}{data.frame with arguments for rnorm to pmap through}

\item{p}{number of predictors}
}
\description{
Get random data
}
