% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{integer_rnorm}
\alias{integer_rnorm}
\title{Integer rnorm}
\usage{
integer_rnorm(n, mean, sd)
}
\arguments{
\item{n}{number of observations passed to rnorm}

\item{mean}{mean passed to rnorm}

\item{sd}{standard deviation passed to rnorm}
}
\description{
Integer rnorm
}
