% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cols.R
\name{map_option_to_function}
\alias{map_option_to_function}
\title{Map option to function}
\usage{
map_option_to_function(x)
}
\arguments{
\item{x}{string ("0", "1", "2" or "3")}
}
\description{
Map option to function
}
