% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(path)
}
\arguments{
\item{path}{path to the xlsx file to load (default is "./data/Rumination.xlsx")}
}
\description{
Read data
}
\details{
Missing path occurs locally in development phase 
and NULL path occurs in the app when no file has been uploaded
}
