% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_with_lag.R
\name{replace_na_with_lag_df}
\alias{replace_na_with_lag_df}
\title{Replace NA with lag (for all numeric columns in data)}
\usage{
replace_na_with_lag_df(data)
}
\arguments{
\item{data}{data.frame}
}
\description{
Replace NA with lag (for all numeric columns in data)
}
