% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_with_reg_line.R
\name{replace_na_with_reg_line}
\alias{replace_na_with_reg_line}
\title{Replace NA with regression line}
\usage{
replace_na_with_reg_line(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
Replace NA with regression line
}
