% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_excel.R
\name{replace_with_na}
\alias{replace_with_na}
\title{Replace with NA}
\usage{
replace_with_na(x)
}
\arguments{
\item{x}{any input}
}
\value{
NA
}
\description{
This is a helper function for `create_template`. It is used to replace cell values with NA
}
