% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{simulate_concave_decrease}
\alias{simulate_concave_decrease}
\title{Simulate concave decrease}
\usage{
simulate_concave_decrease(data)
}
\arguments{
\item{data_pmap}{data.frame with arguments for rnorm to pmap through}
}
\description{
Simulate concave decrease
}
