% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{simulate_convex_decrease}
\alias{simulate_convex_decrease}
\title{Simulate convex decrease}
\usage{
simulate_convex_decrease(data)
}
\arguments{
\item{data_pmap}{data.frame with arguments for rnorm to pmap through}
}
\description{
Simulate convex decrease
}
