% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{simulate_linear_decrease}
\alias{simulate_linear_decrease}
\title{Simulate linear decrease}
\usage{
simulate_linear_decrease(data)
}
\arguments{
\item{data_pmap}{data.frame with arguments for rnorm to pmap through}
}
\description{
Simulate linear decrease
}
