% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{simulate_zero_change}
\alias{simulate_zero_change}
\title{Simulate zero change}
\usage{
simulate_zero_change(data)
}
\arguments{
\item{data_pmap}{data.frame with arguments for rnorm to pmap through}
}
\description{
Simulate zero change
}
