% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{construct_data}
\alias{construct_data}
\title{Construct data}
\usage{
construct_data(seed = 1)
}
\arguments{
\item{seed}{seed}
}
\value{
tibble with columns Deltager, Uge and simulated answers for predictors 1:17
}
\description{
Construct data
}
