% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_random_data}
\alias{get_random_data}
\title{Get random data}
\usage{
get_random_data(data, trend, p)
}
\arguments{
\item{data}{tibble with Uge column}

\item{trend}{string specifying which simulate_data function to use('linear decrease', 'unchanged', 'convex decrease' or 'concave decrease')}

\item{p}{number of predictors}
}
\description{
Get random data
}
