% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_random_vector}
\alias{get_random_vector}
\title{Get random vector}
\usage{
get_random_vector(data_pmap)
}
\arguments{
\item{data_pmap}{data.frame with arguments for rnorm to pmap through}
}
\description{
Get random vector
}
