% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_question.R
\name{plot_question}
\alias{plot_question}
\title{Plot question}
\usage{
plot_question(data = construct_data(p = 17), str_question = "neg_mind")
}
\arguments{
\item{data}{tibble returned from construct_data(p = 17)}

\item{str_question}{string specifying which question to plot 
("neg_mind", "pos_mind", "old_strategy" or "new_strategy")}
}
\value{
plotly object
}
\description{
Plot question
}
\examples{
data <- construct_data(p = 17)
plot_question(data = data, str_question = "neg_mind")

}
